Crate solana_quic_definitions

Definitions related to Solana over QUIC.

Constants

QUIC_CONNECTION_HANDSHAKE_TIMEOUT

QUIC_KEEP_ALIVE

To avoid idle timeout, the QUIC endpoint sends a ping every QUIC_KEEP_ALIVE. This shouldn’t be too low to avoid unnecessary ping traffic.

QUIC_MAX_STAKED_CONCURRENT_STREAMS

QUIC_MAX_STAKED_RECEIVE_WINDOW_RATIO

The receive window for QUIC connection from maximum staked nodes is set to this ratio times solana_packet::PACKET_DATA_SIZE

QUIC_MAX_TIMEOUT

QUIC connection idle timeout. The connection will be closed if there are no activities on it within the timeout window.

QUIC_MAX_UNSTAKED_CONCURRENT_STREAMS

QUIC_MIN_STAKED_CONCURRENT_STREAMS

QUIC_MIN_STAKED_RECEIVE_WINDOW_RATIO

The receive window for QUIC connection from minimum staked nodes is set to this ratio times solana_packet::PACKET_DATA_SIZE

QUIC_PORT_OFFSET

QUIC_SEND_FAIRNESS

QUIC_TOTAL_STAKED_CONCURRENT_STREAMS

QUIC_UNSTAKED_RECEIVE_WINDOW_RATIO

The receive window for QUIC connection from unstaked nodes is set to this ratio times solana_packet::PACKET_DATA_SIZE

Traits

NotifyKeyUpdate